Split OngoingGame and CompletedGame
