+++ src/GameRoom/CompletedGame.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import { PlayerID, Player } from '../api'
+
+import * as PlayersRow from './PlayersRow'
+
+export const CompletedGame: React.VFC<{
+  winnerId: PlayerID
+  players: ReadonlyArray<Player>
+}> = React.memo(({ winnerId, players }) => {
+  return (
+    <>
+      <PlayersRow.CompletedGame winnerId={winnerId} players={players} />
+    </>
+  )
+})
